gml: add module, implement dnd CodeAction
This is a huge commit. Trying to break it up into multiple but it's still going to be a mess inevitably.
So this one adds the gml & gml::dnd modules, and implements the (basic) reading of drag-n-drop CodeActions.
It's not currently parsed beyond reading the data. The implementation is also a bit awkward as you have to pass in a Cursor.
This is due to the size being unknown from gamedata, which is stupid but that's how YoYoGames did it. Thanks.
I need to make an enum eventually of the ID's and of course parse it into an executable action.
I obviously need to make the GML runner first so I'll put this on hold at least until I'm done reading all assets.