reader: extract icon data from exe
This was quite a substantial change to the reader which I've been wanting to do for a while. Previously we were making a lot of unhealthy assumptions about the PE header layout; this commit added a routine which reads the entire header including all COFF sections and extracts the info it needs for later.