build: Add and set the application manifest for Windows EXE This is mainly necessary to use comctl32 (aka 'commctrl') library of version 6 (default is 5), which is used by the original GameMaker and thus required for extensions like Max WinAPI 2 to work properly, because this version contains some new controls. https://docs.microsoft.com/en-us/windows/win32/controls/cookbook-overview https://docs.microsoft.com/en-us/windows-hardware/drivers/print/using-theme-manifests
Until the extension emulation infrastructure appears, it should be considered only as a temporary solution. Ideally,
the gm8emulator should not be linked to the comctl32 at all on Windows. Instead, if needed, it should load e.g. the
Max WinAPI 2 replacement DLL (a reimplementation), which should in its turn be both linked to comctl32 and containing
a manifest to make Windows DLL loader acquire the proper version of it in gm8emulator run-time.

This manifest also marks the application as UAC-familiar (the 'trustInfo' XML section) to disable the registry access
virtualization, so registry functions of the engine API could be properly implemented in future. GameMaker 8.1 has this
as well, while 8.0 does not, due to the older Delphi compiler that was used back in time of its development.
https://docs.microsoft.com/en-us/previous-versions/bb756929(v=msdn.10)

Other useful MSDN links to better understand the manifest format, its abilities and schemas:
https://docs.microsoft.com/en-us/windows/win32/sbscs/application-manifests
https://docs.microsoft.com/en-us/visualstudio/deployment/clickonce-application-manifest?view=vs-2019
https://docs.microsoft.com/en-us/openspecs/sharepoint_protocols/ms-osco/611cc06e-88c6-4a31-a75a-e496c4125ee9

It's important to note that Windows caches manifests of EXE files, so care must be taken when making any changes - you
must ensure while testing that every new build iteration will obtain an unique EXE filename, or rename it by yourself.

Also note that although the documentation states that 'processorArchitecture' attribute is optional and doesn't even
indicate asterisk as one of it's possible values (while it's being used literally everywhere), removing it makes
Windows EXE loader throw an error at application start.