events: Pass global mouse events from GM 5.x
In the partial_demo.gm81 example by @renex, it was noticed that the "An Attractor" and "Destroyer" screens did not respond to mouse clicks in OpenGMK. It turned out that the reason was the version of GameMaker it was originally written for (5.1, which introduced particle support). In the 5.x branch, global mouse events were not differentiated by buttons. Starting with the 6.x branch, this was changed, but the old events were still executed for backwards compatibility, although they were now displayed in the IDE as "<unknown>". But this was not taken into account when the example was ported to newer versions of GM.